Fixed argument type for precondition.

+++ include/dangerfarm_contact/util/socket.h
@@ -54,7 +54,7 @@ int DANGERFARM_CONTACT_SYM(socket_write_uint64)(int s, uint64_t val);
 /* preconditions. */
 MODEL_CONTRACT_PRECONDITIONS_BEGIN(
     DANGERFARM_CONTACT_SYM(socket_write_uint64),
-    int s, uint32_t val)
+    int s, uint64_t val)
         MODEL_ASSERT(prop_is_open_fd(s));
 MODEL_CONTRACT_PRECONDITIONS_END(DANGERFARM_CONTACT_SYM(socket_write_uint64))
